Beefed up the help file. We now print out available Filters (and their attributes and methods) and ResultSets automatically so long as they follow some conventions i.e. Filter (resp. ResultSet) in their name, private members prefixed with '_' etc.
Added a lot more examples to filterdemo.py

kobject: Added function for returning class functions, methods, and other attributes.
Fixed a bug with escaping the wildcard % in a LIKE statement when there are existing string formatting.
Fixed a bug where self.paramsoffset was not being reset before each application and so broke the assertion in Filter.
Turn off some printing in ResultSet which should only be turned on for debugging.


git-svn-id: file:///kortemmelab/shared/commonSVN/ddg/trunk@577 fc5780ad-6544-46e6-b60b-86c75d3988a5