Adding two versions of the same function as a lesson learned.
One version is seemingly optimized for memory allocation however it is much slower. Within that version, a for-loop using xrange out-performs a while loop.
Lesson learned: do not try to program Python like C/C++. Simple operations in Python are not optimized by the compiler so it is usually quicker to run C/C++-bound functions.