Abacus graph update. The graph now looks nice on a range of widths and lengths.
Fixed the formatting to scale with the number of points. Some problems were simply the order of the commands.
Set a minimum width and height for the graph.
Tested the graph with small numbers of labels (1 label) and changed the code to suit.
Added practical assertions (number of labels = number of datapoints per line).