Adding get_rcsb_record to get the associated RCSB record from a derived structure (we assume a tree hierarchy i.e. if a design is based on another design then the second design is referred to in the DerivedFrom field etc. until the first design is derived from an RCSB file).
Adding _check_derived_record_against_rcsb_record to do basic tests (the chain mapping respects chain molecule types).

Allowing molecules to be added for derived structures. We copy the molecule information from the original PDB which is useful for when the derived file is missing the headers or if chains were renamed and the headers left untouched.